C#: add "regular" .NET 4.6 project too (TestCsRegular).
I would have used one of them "multi-targeting" thingies in one .csproj, but VSMac does not support them, i.e. no UI to pick which one to target from the IDE.

So a separate .csproj it is. Which of course is complicated in something called project.json (whatever that is, I've no idea), and needs a custom <BaseIntermediateOutputPath> tag in the new .csproj. See https://github.com/NuGet/Home/issues/5126

Computers were a mistake!

Anyhoo, numbers in Mray/s on Mac (see issue #3):
- dotnet (2.1.4): 13.0
- mono (5.8.1): 5.5
- mono -O=float32: 6.0
- mono --llvm: 7.8
- mono --llvm -O=float32: 10.5
- MONO_INLINELIMIT=100 mono --llvm -O=float32: 12.7